feat: Add doven-kalender application orchestration and event processing
- Introduced main application logic in `__init__.py` for orchestrating event extraction, parsing, and description generation.
- Implemented event parsing with LLM in `parse_event_with_llm` and description generation in `generate_description`.
- Added calendar event extraction from Google Calendar API in `extractor/kalender.py`.
- Created LLM integration for Google, OpenAI, and Mistral in `generator/generator.py`.
- Added emoji assignment functionality in `parser/get_emojis.py`.
- Established structured data models using Pydantic in `parser/parser.py`.
- Created CLI entry points for running the application and processing CSV to JSON conversion.
- Updated project configuration in `pyproject.toml` for packaging and dependencies.
- Added tests for event extraction and emoji assignment functionalities.
- Removed obsolete files and cleaned up the project structure.
