+++ .gitignore
@@ -384,3 +384,4 @@ $RECYCLE.BIN/
 *.lnk
 
 # End of https://www.toptal.com/developers/gitignore/api/python,visualstudiocode,pycharm,linux,windows,macos
+AGENTS.md
