+++ doven_kalender/__main__.py
@@ -1,0 +1,7 @@
+"""Allow running the app with `python -m doven_kalender`."""
+
+from . import main
+
+
+if __name__ == "__main__":
+    raise SystemExit(main())
