+++ doven_kalender/extractor/__init__.py
@@ -1,4 +1,4 @@
-"""Extractor package for doven-kalender."""
+"""Extractor package for doven-kalender (packaged namespace)."""
 
 __all__ = ["extract_csv_from_odt", "kalender"]
