+++ doven_kalender/generator/poster.py
@@ -1,0 +1,4 @@
+
+"""Poster helpers (packaged namespace)."""
+
+__all__ = []
