+++ doven_kalender/parser/convert_csv_to_json.py
@@ -30,8 +30,6 @@ DATE_PATTERN = re.compile(r"(\d+)\s*\.?\s*([^\W\d_]+)", re.UNICODE)
 
 
 def extract_date_parts(date_text: str) -> tuple[int, str] | None:
-    """Extract the day number and month token from a date string."""
-
     match = DATE_PATTERN.search(date_text.strip())
     if match is None:
         return None
@@ -40,8 +38,6 @@ def extract_date_parts(date_text: str) -> tuple[int, str] | None:
 
 
 def normalize_month_name(month_name: str) -> str | None:
-    """Normalize a Danish month name to its three-letter abbreviation."""
-
     normalized_month = month_name.strip().lower()
     if normalized_month in MONTH_ABBRS:
         return normalized_month
@@ -50,8 +46,6 @@ def normalize_month_name(month_name: str) -> str | None:
 
 
 def normalize_date(date_text: str) -> str:
-    """Normalize a date like '8. juli' to '8. jul'."""
-
     date_parts = extract_date_parts(date_text)
     if date_parts is None:
         return date_text.strip()
@@ -65,8 +59,6 @@ def normalize_date(date_text: str) -> str:
 
 
 def normalize_target_month(month_name: str) -> str:
-    """Normalize a target month to the abbreviation used in the JSON output."""
-
     normalized_month = normalize_month_name(month_name)
     if normalized_month is not None:
         return normalized_month
@@ -75,20 +67,14 @@ def normalize_target_month(month_name: str) -> str:
 
 
 def classify_month(month_abbr: str, left_month: str, right_month: str) -> str | None:
-    """Return the output side for a normalized month, or None if it does not match."""
-
     if month_abbr == left_month:
         return "left"
-
     if month_abbr == right_month:
         return "right"
-
     return None
 
 
 def build_entry(row: dict[str, str]) -> dict[str, str | None]:
-    """Build the JSON entry for a single CSV row."""
-
     return {
         "date": normalize_date(row.get("Dato", "")),
         "emoji": None,
@@ -103,15 +89,6 @@ def convert_csv_to_json(
     left_month: str,
     right_month: str,
 ) -> None:
-    """Convert a CSV file to the left/right JSON structure used by the poster.
-
-    Args:
-        csv_file: Path to the input CSV file.
-        json_file: Path to the output JSON file.
-        left_month: Month assigned to the ``left`` side.
-        right_month: Month assigned to the ``right`` side.
-    """
-
     left_month_abbr = normalize_target_month(left_month)
     right_month_abbr = normalize_target_month(right_month)
     if left_month_abbr == right_month_abbr:
@@ -142,8 +119,6 @@ def convert_csv_to_json(
 
 
 def parse_args(argv: Sequence[str]) -> argparse.Namespace:
-    """Parse command line arguments for the CSV conversion script."""
-
     parser = argparse.ArgumentParser(description=__doc__)
     parser.add_argument("input_csv", type=Path, help="Path to the input CSV file")
     parser.add_argument("output_json", type=Path, help="Path to the output JSON file")
@@ -153,8 +128,6 @@ def parse_args(argv: Sequence[str]) -> argparse.Namespace:
 
 
 def main(argv: Sequence[str] | None = None) -> int:
-    """Run the CSV to JSON conversion CLI."""
-
     args = parse_args(sys.argv[1:] if argv is None else argv)
     convert_csv_to_json(args.input_csv, args.output_json, args.left_month, args.right_month)
     return 0
