+++ doven_kalender/parser/parser.py
@@ -6,12 +6,10 @@ from typing import Any
 
 from pydantic import BaseModel, Field
 
-from generator.generator import GeneralLLM
+from ..generator.generator import GeneralLLM
 
 
 class ParsedEvent(BaseModel):
-    """Structured event representation used by downstream generators/integrators."""
-
     title: str = Field(description="Event title")
     description: str = Field(description="Event description")
     start_time: str = Field(description="Start timestamp in ISO-8601 format")
@@ -19,26 +17,10 @@ class ParsedEvent(BaseModel):
 
 
 class EventParser:
-    """Parser service that uses a generic LLM to extract structured event fields."""
-
     def __init__(self, llm: GeneralLLM):
-        """Initialize parser with the shared LLM interface.
-
-        Args:
-            llm: Concrete LLM implementation used for parsing.
-        """
         self._llm = llm
 
     def parse_event(self, event: dict[str, Any], response_model: type[BaseModel] = ParsedEvent) -> BaseModel:
-        """Parse an extracted event dict into a structured Pydantic model.
-
-        Args:
-            event: Raw event object from the extractor.
-            response_model: Structured output schema. Defaults to `ParsedEvent`.
-
-        Returns:
-            Parsed event as the requested Pydantic model.
-        """
         prompt = (
             "Parse this calendar event into structured data with normalized title, "
             "description, and ISO timestamps:\n"
