+++ pyproject.toml
@@ -13,3 +13,10 @@ dependencies = [
 [project.scripts]
 # Console entrypoint for running the CLI: `doven-kalender`
 doven-kalender = "doven_kalender:main"
+
+[build-system]
+requires = ["setuptools>=61.0", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+packages = { find = { where = ["."], include = ["doven_kalender*"] } }
